Formatted Code

+++ contracts/swaps/connectors/SwapsImplUniswapV3_ETH.sol
@@ -15,7 +15,7 @@ import "../../mixins/Path.sol";
 contract SwapsImplUniswapV3_ETH is State, ISwapsImpl {
     using SafeERC20 for IERC20;
     using Path for bytes;
-	using BytesLib for bytes;
+    using BytesLib for bytes;
     address public uniswapSwapRouter;
     address public uniswapQuoteContract;
 
@@ -146,65 +146,91 @@ contract SwapsImplUniswapV3_ETH is State, ISwapsImpl {
         returns (uint256 sourceTokenAmountUsed, uint256 destTokenAmountReceived)
     {
         if (requiredDestTokenAmount != 0) {
-			(bytes[] memory routes,uint[] memory amountsIn, uint[] memory amountsOut) = abi.decode(payload,(bytes[], uint[], uint[]));
-			require(routes.length==amounts.length);
-			bytes[] memory encodedTXs = new bytes[](amounts.length);
-			uint totalAmountsOut = 0;
-			uint totalAmountsInMax = 0;
-			for(uint x=0;x<amountsIn.length;x++){
-				(, address tokenIn, ) = routes[x].decodeFirstPool();
-				require(tokenIn == sourceTokenAddress, "improper route");
-				address tokenOut = routes[x].toAddress(routes[x].length-20);
-				require(tokenOut == destTokenAddress, "improper destination");
-				IUniswapV3SwapRouter.ExactOutputParams
-					memory swapParams = IUniswapV3SwapRouter.ExactOutputParams({
-						path: routes[x],
-						recipient: receiverAddress,
-						deadline: block.timestamp,
-						amountOut: amountsOut[x],
-						amountInMaximum: amountsIn[x]
-					});
-					totalAmountsOut = totalAmountsOut + amountsOut[x];
-					totalAmountsInMax = totalAmountsInMax + amountsIn[x];
-				encodedTXs[x] = abi.encodeWithSelector(IUniswapV3SwapRouter(uniswapSwapRouter).exactOutput.selector,swapParams);
-			}
-			require(totalAmountsOut==requiredDestTokenAmount&&totalAmountsInMax<=maxSourceTokenAmount);
-			bytes[] memory trueAmountsIn = IUniswapV3SwapRouter(uniswapSwapRouter).multicall(encodedTXs);
-			uint totaledAmountIn = 0;
-			for(uint x=0;x<trueAmountsIn.length;x++){
-				totaledAmountIn = totaledAmountIn + abi.decode(trueAmountsIn[x],(uint));
-			}
-			sourceTokenAmountUsed = totaledAmountIn;
-			destTokenAmountReceived = requiredDestTokenAmount;
+            (
+                bytes[] memory routes,
+                uint256[] memory amountsIn,
+                uint256[] memory amountsOut
+            ) = abi.decode(payload, (bytes[], uint256[], uint256[]));
+            require(routes.length == amounts.length);
+            bytes[] memory encodedTXs = new bytes[](amounts.length);
+            uint256 totalAmountsOut = 0;
+            uint256 totalAmountsInMax = 0;
+            for (uint256 x = 0; x < amountsIn.length; x++) {
+                (, address tokenIn, ) = routes[x].decodeFirstPool();
+                require(tokenIn == sourceTokenAddress, "improper route");
+                address tokenOut = routes[x].toAddress(routes[x].length - 20);
+                require(tokenOut == destTokenAddress, "improper destination");
+                IUniswapV3SwapRouter.ExactOutputParams
+                    memory swapParams = IUniswapV3SwapRouter.ExactOutputParams({
+                        path: routes[x],
+                        recipient: receiverAddress,
+                        deadline: block.timestamp,
+                        amountOut: amountsOut[x],
+                        amountInMaximum: amountsIn[x]
+                    });
+                totalAmountsOut = totalAmountsOut + amountsOut[x];
+                totalAmountsInMax = totalAmountsInMax + amountsIn[x];
+                encodedTXs[x] = abi.encodeWithSelector(
+                    IUniswapV3SwapRouter(uniswapSwapRouter)
+                        .exactOutput
+                        .selector,
+                    swapParams
+                );
+            }
+            require(
+                totalAmountsOut == requiredDestTokenAmount &&
+                    totalAmountsInMax <= maxSourceTokenAmount
+            );
+            bytes[] memory trueAmountsIn = IUniswapV3SwapRouter(
+                uniswapSwapRouter
+            ).multicall(encodedTXs);
+            uint256 totaledAmountIn = 0;
+            for (uint256 x = 0; x < trueAmountsIn.length; x++) {
+                totaledAmountIn =
+                    totaledAmountIn +
+                    abi.decode(trueAmountsIn[x], (uint256));
+            }
+            sourceTokenAmountUsed = totaledAmountIn;
+            destTokenAmountReceived = requiredDestTokenAmount;
         } else {
-			(bytes[] memory routes,uint[] memory amounts) = abi.decode(payload,(bytes[], uint[]));
-			require(routes.length==amounts.length);
-			bytes[] memory encodedTXs = new bytes[](amounts.length);
-			uint totalAmounts = 0;
-			for(uint x=0;x<amounts.length;x++){
-				(, address tokenIn, ) = routes[x].decodeFirstPool();
-				require(tokenIn == sourceTokenAddress, "improper route");
-				address tokenOut = routes[x].toAddress(routes[x].length-20);
-				require(tokenOut == destTokenAddress, "improper destination");
-				IUniswapV3SwapRouter.ExactInputParams
-					memory swapParams = IUniswapV3SwapRouter.ExactInputParams({
-						path: routes[x],
-						recipient: receiverAddress,
-						deadline: block.timestamp,
-						amountIn: amounts[x],
-						amountOutMinimum: 1
-					});
-					totalAmounts = totalAmounts + amounts[x];
-				encodedTXs[x] = abi.encodeWithSelector(IUniswapV3SwapRouter(uniswapSwapRouter).exactInput.selector,swapParams);
-			}
-			sourceTokenAmountUsed = totalAmounts;
-			require(totalAmounts<=maxSourceTokenAmount);
-			bytes[] memory trueAmountsOut = IUniswapV3SwapRouter(uniswapSwapRouter).multicall(encodedTXs);
-			uint totaledAmountOut = 0;
-			for(uint x=0;x<trueAmountsOut.length;x++){
-				totaledAmountOut = totaledAmountOut + abi.decode(trueAmountsOut[x],(uint));
-			}
-			destTokenAmountReceived = totaledAmountOut;
+            (bytes[] memory routes, uint256[] memory amounts) = abi.decode(
+                payload,
+                (bytes[], uint256[])
+            );
+            require(routes.length == amounts.length);
+            bytes[] memory encodedTXs = new bytes[](amounts.length);
+            uint256 totalAmounts = 0;
+            for (uint256 x = 0; x < amounts.length; x++) {
+                (, address tokenIn, ) = routes[x].decodeFirstPool();
+                require(tokenIn == sourceTokenAddress, "improper route");
+                address tokenOut = routes[x].toAddress(routes[x].length - 20);
+                require(tokenOut == destTokenAddress, "improper destination");
+                IUniswapV3SwapRouter.ExactInputParams
+                    memory swapParams = IUniswapV3SwapRouter.ExactInputParams({
+                        path: routes[x],
+                        recipient: receiverAddress,
+                        deadline: block.timestamp,
+                        amountIn: amounts[x],
+                        amountOutMinimum: 1
+                    });
+                totalAmounts = totalAmounts + amounts[x];
+                encodedTXs[x] = abi.encodeWithSelector(
+                    IUniswapV3SwapRouter(uniswapSwapRouter).exactInput.selector,
+                    swapParams
+                );
+            }
+            sourceTokenAmountUsed = totalAmounts;
+            require(totalAmounts <= maxSourceTokenAmount);
+            bytes[] memory trueAmountsOut = IUniswapV3SwapRouter(
+                uniswapSwapRouter
+            ).multicall(encodedTXs);
+            uint256 totaledAmountOut = 0;
+            for (uint256 x = 0; x < trueAmountsOut.length; x++) {
+                totaledAmountOut =
+                    totaledAmountOut +
+                    abi.decode(trueAmountsOut[x], (uint256));
+            }
+            destTokenAmountReceived = totaledAmountOut;
         }
     }
 }
